requirements.txt
yfinance
pandas
requests
ta
python-telegram-bot[job-queue]==21.10
Flask==3.0.3
